Fix Blueprint type error: change NodeId from uint16 to int32
uint16 is not supported by Blueprint reflection. Changed NodeId to int32
(Blueprint-compatible) and updated BEHAVIAC_INVALID_NODE_ID constant
accordingly. Removed unnecessary static_cast in LoadFromProperties.

+++ BehaviacPlugin/Source/BehaviacRuntime/Private/BehaviorTree/BehaviacBehaviorNode.cpp
@@ -19,7 +19,7 @@ void UBehaviacBehaviorNode::LoadFromProperties(int32 Version, const FString& InA
 	{
 		if (Prop.Name == TEXT("Id"))
 		{
-			NodeId = static_cast<uint16>(FCString::Atoi(*Prop.Value));
+			NodeId = FCString::Atoi(*Prop.Value);
 		}
 	}
 }
